requirements.txt
streamlit>=1.28.0
google-generativeai>=0.3.0
python-dateutil>=2.8.2
python-dotenv